+++ CepGenAddOns/RivetWrapper/YODAHistsHandler.cpp
@@ -148,16 +148,16 @@ namespace cepgen {
   void YODAHistsHandler<T>::operator<<(const Event& ev) {
     //--- increment the corresponding histograms
     for (auto& h_var : hists1d_)
-      h_var.second.fillBin(browser_.get(ev, h_var.first), (double)cross_section_);
+      h_var.second.fillBin(browser_.get(ev, h_var.first), cross_section_);
     for (auto& h_var : hists2d_)
-      h_var.second.fillBin(browser_.get(ev, h_var.first[0]), browser_.get(ev, h_var.first[1]), (double)cross_section_);
+      h_var.second.fillBin(browser_.get(ev, h_var.first[0]), browser_.get(ev, h_var.first[1]), cross_section_);
     for (auto& h_var : profiles1d_)
-      h_var.second.fill(browser_.get(ev, h_var.first[0]), browser_.get(ev, h_var.first[1]), (double)cross_section_);
+      h_var.second.fill(browser_.get(ev, h_var.first[0]), browser_.get(ev, h_var.first[1]), cross_section_);
     for (auto& h_var : profiles2d_)
       h_var.second.fill(browser_.get(ev, h_var.first[0]),
                         browser_.get(ev, h_var.first[1]),
                         browser_.get(ev, h_var.first[2]),
-                        (double)cross_section_);
+                        cross_section_);
     weight_cnt_.fill(ev.weight);
   }
